// src/api/generated/models/pageResponseAdminUserResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ASMS API
 * Assignment Management System - REST API
 * OpenAPI spec version: v1
 */
import type { AdminUserResponse } from './adminUserResponse'

export interface PageResponseAdminUserResponse {
  items: AdminUserResponse[]
  page: number
  size: number
  totalElements: number
  totalPages: number
}